Modified coulomb_c basin-scale inversion
This commit changes some coulomb_c options as follows:
* changes the basin-scale inversion options for which_ho_coulomb_c
* removes the option which_ho_coulomb_c_relax
* adds the option elevation_based_coulomb_c.

The new set of which_ho_coulomb_c options is
  HO_COULOMB_C_CONSTANT = 0
  HO_COULOMB_C_INVERSION = 1
  HO_COULOMB_C_EXTERNAL = 2
  HO_COULOMB_C_INVERSION_BASIN = 3
  HO_COULOMB_C_EXTERNAL_BASIN = 4

Options 0, 1 and 2 are as before.

Option 3 is modified:
* In the old version, the inversion was for a single value of Cc in each basin.
  A single value turns out to be unrealistic.
  This value is too low in high-elevation regions and/or too low in low-elevation regions.
* In the new version, Cc is assumed to be elevation-dependent:
  Cc = coulomb_c_hi for z >= bed_hi
  Cc = coulomb_c_lo for z <= bed_lo
  log(Cc) is interpolated linearly between log(coulomb_c_lo) and log(coulomb_c_hi) for z between bed_lo and bed_hi
This is the same elevation-based formulation CISM has used before, except that log(Cc)
 instead of Cc is now linearly interpolated.
Given this elevation dependence, the inversion is for coulomb_c_lo in each basin.
This makes it possible to invert for a low value in some basins (e.g., the Siple coast)
and higher values in other basins. Unless Cc_lo is small, the Siple ice tends to be too thick.
In the new code, both coulomb_c_lo and coulomb_c_hi are 2D fields, although only
 coulomb_c_lo is spatially varying; coulomb_c_hi is set to a constant everywhere.
The elevation target is the mean elevation of marine-grounded_ice at initialization.

Option 4 is new. It is related to option 3 in the same way option 2 is related to option 1.
If doing basin-scale inversion, we use option 3 for the spin-up, and then run forward
with option 4, reading the saved value of coulomb_c_lo from the restart file.

I removed option which_ho_coulomb_c_relax, which has not been very useful.
When inverting locally for coulomb_c, the relaxation is now toward coulomb_c_const
rather than an elevation-based range of options.

I added the logical config option elevation_based_coulomb_c. The default is F.
This option must be set to F for which_ho_coulomb_c = 1 or 2, and to T for
which_ho_coulomb_c = 3 or 4.

To run with fixed elevation-based Cc, the user sets which_ho_coulomb_c = 0
and elevation_based_coulomb_c = T. This results in both coulomb_c_hi and coulomb_c_lo being constants,
which are 0.5 and 0.1 by default but can be set in the config file.

The defaults for bed_hi and bed_lo are 0 m and -500 m. The default bed_hi is lower than I've used before,
 to allow higher Cc for all ice grounded above sea level.
These values can be set by the user in the config file.

I tried some AIS spin-ups with the new which_ho_coulomb_c = 3, along with active subglacial hydrology.
Results look reasonable overall, but there are some large thickness biases, e.g.:
* PIG is too thick and Thwaites is too thin
* The Siple coast ice is too thick near the GL and too thin farther upstream.
* The Transantarctic Mountains and AP are too thick.
The hope is that Thwaites will improve with a new target thickness dataset, which would
 improve PIG as well (since the Thwaites errors leak into the PIG basin).
The Siple coast might improve with more work on the hydrology.
The TA and AP might improve with either hydrology changes or finer resolution.

Other changes:

* Replaced the parallel_is_nonzero function with parallel_is_zero, and added
  several calls to this function to simplify some logic

* Fixed the SMB units for one of the calls to check_fill_values (to check dthck_dt_obs)

* Reduced the default for bwat_threshold to 1.e-3

I confirmed that this commit is BFB for several Antarctic and glacier applications
that do not use the new options.