Changed the function relating N to subglacial bwat
A key link between the active subglacial hydrology and the ice dynamics
is the function that relates h = bwat (the subglacial water depth) to the
effective pressure N.

The previous function was N/overburden = 1 - h/(h + h0),
where h0 = bwat_threshold.
This function asymptotes to zero very slowly as h increases relative to h0.

The new function is N/overburden = max[delta, exp(-h/h0)], where h0 = bwat_threshold
is now an e-folding scale, and delta ~ 0.02.
This function asymptotes to delta quickly as h increases relative to h0,
but not so quickly as to make the dynamics unstable, at least in tests to date.