Added zatm coordinate for 3D atmospheric forcing
This commit adds zatm as a dimension for atmosphere input in multiple vertical layers.
This dimension is useful for reading and writing surface mass balance, temperature,
or other fields in multiple elevation classes (e.g., as simulated by CESM).
Previously, there was a similar variable called nlev_smb, but it didn't contain
all the necessary code for I/O. I changed the name to zatm in analogy to zocn.
The implementation of zatm is very similar to that of zocn.

To read atmospheric data in multiple ECs and downscale it to the ice surface,
the user should set smb_input_function = artm_input_function = 2.
The number of levels nzatm and the levels (e.g., 100, 300, ...) in units of meters
can be specified in the config file in a section called [grid_atm].
If they are not specified, CISM will default to the 10 standard CESM levels.

The downscaling takes place in subroutine prepare_climate_forcing in glide_mass_balance.F90.
There are changes in other files including glide.F90, glide_types.F90, glide_setup.F90,
glide_vars.def, and generate_ncvars.py to support zatm.
I also fixed a logic error in subroutine glissade_vertical_interpolate.