Added a heat flux for refreezing subglacial meltwater
This commit applies to the flux-routing subglacial hydrology.
It aims to treat subglacial hydrology more realistically by focusing the flow along thawed beds.
When meltwater flows into a cell with a frozen bed, some or all of the water can refreeze in place
instead of flowing downstream.

The heat flux associated with refreezing is called bhdyroflx.
This flux enters the basal energy balance along with the geothermal and frictional heat fluxes.
If bhydroflx is large enough, it can thaw the bed, allowing meltwater to flow more freely.

There is one major change in the flux-routing algorithm to focus the flow along thawed beds.
For each cell we define a ramp function based on the quantity delta_Tb = bpmp - btemp,
i.e. the difference between the PMP temperature and the actual bed temperature.
Here, btemp_scale is a temperature scale with units of deg C.
When basal water flows into a cell with delta_Tb > btemp_scale, the water is refrozen in place.
If 0 < delta_Tb < btemp_scale, a fraction delta_Tb/btemp_scale is refrozen in place.
The associated heat flux bhydroflx is passed to glissade_therm_driver.

I added bhydroflx to glide_vars.def. It is a restart variable so that it's available
during the next timestep (similar to the treatment of bfricflx).

This change creates a potential mechanism for turning ice streams on and off.
The goal is to simulate transient behavior for glaciers like Ice Stream C (aka Kamb).
Preliminary tests showed some weakening of Ice Stream C, but more testing is needed.

This commit also fixes an exact restart issue when inverting for basin-based coulomb_c_lo.