More work toward reproducible sums
This commit continues the work toward a reproducible sum capability:
an option to compute sums such that results on different numbers
of processors are BFB.

The basic capability is now working. In the various parallel_global_sum routines
(with interfaces parallel_global_sum, parallel_global_sum_stagger and
parallel_global_sum_patch), there is an option to call subroutine
parallel_reduce_reprosum in lieu of parallel_reduce sum.
Subroutine parallel_reduce_reprosum calls cism_reproduce_calc, which
converts dp variables to i8 and does the sums.

I verified that sums are now reproducible for some test problems, including
a Greenland configuration similar to what we are using for CESM3.
One issue was that the velocity solver was using local copies of
variables xVertex and yVertex for its node geometry, which led to a dependence
on processor count. Computing these variables based on the global coordinates
model%general%x0 and model%general%y0 solves the problem, although it leads
to slightly larger roundoff errors for some calculations.

I fixed another subtle issue by rewriting two equations of this form:
  c1 = c1 + a*b
  c2 = c2 + a*b
with equations of this form:
  f = a*b
  c1 = c1 + f
  c2 = c2 + f
I don't know why this was necessary.

Some results still depend on processor count: for instance, the scheme that routes
subglacial water downhill to neighboring cells. I will work on that in a future commit.

Other changes:

* Fixed a bug in the way the rmse for surface velocity is computed.

* Added a subroutine called double_to_binary in glimmer_utils.
  Given an input double-precision floating-point number, the output is a
  character string corresponding to the 64-bit internal representation of that number
  (1 bit for sign, 11 for exponent, and 52 for fraction).
  This is useful for debugging.

* Added two 'write_array_to_file' subroutines (real8_2d and real8_3d) in glissade_utils.
  These subroutines take an input array of dp variables, convert each variable
  to a character string corresponding to the 64-bit binary representation,
  and write that string to a file. This can be useful for debugging if we suspect
  that the values of some array elements depend on processor count, but we
  don't know which elements. The real8_3d version has an optional argument to make
  it work for either (k,i,j) arrays or (i,j,k) arrays.